Refactor proof of concept and set up formal structure.
