+++ senior-sage/package.json
@@ -17,7 +17,9 @@
     "framer-motion": "^11.0.20",
     "ollama": "^0.5.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^1",
